refactoring: move views in separate folder
